RDR-010: extract shared PyramidKeyDecoder — dedup pyramidFromKey (Luciferase-3y1)
The root-to-key pyramid descent was duplicated verbatim (~50 lines) in PyramidIndex.pyramidFromKey
(private static) and PyramidSubdivisionStrategy.pyramidFromKey (the strategy has no back-reference to
the index). Extracted to a shared package-private PyramidKeyDecoder.pyramidFromKey; both call sites now
delegate, so a future SFC-encoding/descent change can't silently diverge between them.

Pure refactor, no behavior change — verified by the green pyramid suite (188 tests, incl. PyramidIndex
parity, codec, subdivision-strategy, ray/plane traversal, neighbor detection).

Luciferase-3y1
